Don't add 1 to the width of .dwwr.
This causes an extra pixel of space between .dwwr and .dwc.  If the contents
are centered in .dwwr, this also results in them being off-center by a pixel,
due to the difference in width being odd.  The extra space between the
boxes may also not be wanted by the style.