Only bind event listeners on document when needed.
Bind mousemove and mouseup only when actually dragging.  Previously,
these events were bound at initialization, and left in place even when
no scroller was open.  This causes event overhead for any page that loads
the script.  Instead, bind these listeners only when actually dragging.

The "if(move)" checks in the event handlers can be removed, but to keep
the patch small and readable I haven't done that.